Separate files commands into own file
