+++ internal/system/system.go
@@ -1,11 +1,8 @@
 package system
 
 import (
-	"errors"
 	"fmt"
-	"os"
 	"os/exec"
-	"strings"
 )
 
 type System interface {
@@ -13,10 +10,6 @@ type System interface {
 	RequireCommand(command string) error
 	// ExecCommand executes a system terminal command
 	ExecCommand(name string, arg ...string) *exec.Cmd
-	// RequireFilesInWd requires that the files exist in the current working directory, or errors if they don't
-	RequireFilesInWd(filenames ...string) error
-	// RequireDir requires that a directory exists
-	RequireDir(path string) error
 }
 
 type DefaultSystem struct {
@@ -38,45 +31,4 @@ func (s *DefaultSystem) RequireCommand(command string) error {
 
 func (s *DefaultSystem) ExecCommand(name string, arg ...string) *exec.Cmd {
 	return s.stdlib.ExecCommand(name, arg...)
-}
-
-func (s *DefaultSystem) RequireFilesInWd(filenames ...string) error {
-	if len(filenames) == 0 {
-		return nil
-	}
-
-	cwd, err := s.stdlib.Getwd()
-	if err != nil {
-		return fmt.Errorf("failed to get current working directory: %w", err)
-	}
-
-	var missingFiles []string
-	for _, file := range filenames {
-		path := fmt.Sprintf("%s/%s", cwd, file)
-		if _, err := s.stdlib.Stat(path); err != nil {
-			if errors.Is(err, os.ErrNotExist) {
-				missingFiles = append(missingFiles, file)
-			} else {
-				// Some other error returned (permission denied, etc.)
-				return fmt.Errorf("failed to check file %s: %w", path, err)
-			}
-		}
-	}
-
-	if len(missingFiles) > 0 {
-		return fmt.Errorf("required filenames not found: %s", strings.Join(missingFiles, ", "))
-	}
-
-	return nil
-}
-
-func (s *DefaultSystem) RequireDir(path string) error {
-	if stat, err := s.stdlib.Stat(path); err != nil && errors.Is(err, os.ErrNotExist) {
-		return fmt.Errorf("required directory not found: %s", path)
-	} else if err != nil {
-		return fmt.Errorf("failed to check file %s: %w", path, err)
-	} else if !stat.IsDir() {
-		return fmt.Errorf("source path %s is not a directory", path)
-	}
-	return nil
 }
